Added the laser line!

+++ suddendeath/Assets/LaserLineBehavior.cs
@@ -3,16 +3,18 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class LaserLineBehavior : MonoBehaviour {
-    public int speed;
-    public bool isVertical;
+    private GameOptions go;
+    private float speed;
     public Rigidbody2D rb2d;
-    
-    public Vector2 direction;
+    private Vector2 direction;
 
 	// Use this for initialization
 	void Start () {
-		
-	}
+        go = Globals.Instance.GameManager.gameOptions;
+        direction = new Vector2(1, 0);
+        speed = Random.Range(go.laserLineMinSpeed, go.laserLineMaxSpeed);
+        if (Random.value < 0.5f) speed = -speed;
+    }
 	
 	// Update is called once per frame
 	void Update () {
@@ -21,16 +23,14 @@ public class LaserLineBehavior : MonoBehaviour {
 
     private void OnTriggerEnter2D(Collider2D collider)
     {
-        WallBehavior wb = collider.gameObject.GetComponent<WallBehavior>();
-        
-        if (wb != null)
+        if (collider.gameObject.GetComponent<WallBehavior>() != null)
         {
             speed = -speed;
         } else
         {
             PlayerController pc = collider.gameObject.GetComponent<PlayerController>();
 
-            if (pc != null)
+            if (pc != null && !pc.GetComponent<PlayerInput>().IsDashing)
             {
                 pc.Kill(PlayerController.KillType.Dissolve);
             }
